Simplify JSON sanitizing
