+++ rummage/lib/util/__init__.py
@@ -7,7 +7,7 @@ import os
 import subprocess
 from itertools import groupby
 from encodings.aliases import aliases
-from .file_strip.json import sanitize_json
+from .sanitize_json import sanitize
 from urllib.request import url2pathname
 from urllib.parse import urlparse
 import html
@@ -188,7 +188,7 @@ def read_json(filename):
 
     try:
         with codecs.open(filename, "r", encoding='utf-8') as f:
-            content = sanitize_json(f.read(), True)
+            content = sanitize(f.read(), True)
         obj = json.loads(content)
     except Exception:
         obj = None
