--- rummage/lib/util/file_strip/__init__.py
@@ -1,1 +1,0 @@
-"""File Strip."""
